script_interface: Move CTP to math submodule
