+++ src/script_interface/CMakeLists.txt
@@ -37,6 +37,7 @@ add_subdirectory(interactions)
 add_subdirectory(lbboundaries)
 add_subdirectory(lees_edwards)
 add_subdirectory(magnetostatics)
+add_subdirectory(math)
 add_subdirectory(mpiio)
 add_subdirectory(observables)
 add_subdirectory(pair_criteria)
