+++ src/script_interface/math/CMakeLists.txt
@@ -1,0 +1,2 @@
+target_sources(Espresso_script_interface
+               PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/initialize.cpp)
