+++ src/script_interface/math/initialize.hpp
@@ -1,0 +1,35 @@
+/*
+ * Copyright (C) 2022 The ESPResSo project
+ *
+ * This file is part of ESPResSo.
+ *
+ * ESPResSo is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * ESPResSo is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef ESPRESSO_SRC_SCRIPT_INTERFACE_MATH_INITIALIZE_HPP
+#define ESPRESSO_SRC_SCRIPT_INTERFACE_MATH_INITIALIZE_HPP
+
+#include <script_interface/ObjectHandle.hpp>
+
+#include <utils/Factory.hpp>
+
+namespace ScriptInterface {
+namespace Math {
+
+void initialize(Utils::Factory<ObjectHandle> *om);
+
+} // namespace Math
+} // namespace ScriptInterface
+
+#endif
