+++ src/script_interface/observables/CylindricalPidProfileObservable.hpp
@@ -27,7 +27,8 @@
 #include "Observable.hpp"
 #include "core/observables/CylindricalPidProfileObservable.hpp"
 
-#include <script_interface/CylindricalTransformationParameters.hpp>
+#include "script_interface/math/CylindricalTransformationParameters.hpp"
+
 #include <utils/constants.hpp>
 
 #include <boost/range/algorithm.hpp>
@@ -141,7 +142,7 @@ public:
 
 private:
   std::shared_ptr<CoreObs> m_observable;
-  std::shared_ptr<CylindricalTransformationParameters> m_transform_params;
+  std::shared_ptr<Math::CylindricalTransformationParameters> m_transform_params;
 };
 
 } /* namespace Observables */
